move abi encoding/decoding to foundation
