+++ yarn-project/acir-simulator/src/index.ts
@@ -1,5 +1,4 @@
 export * from './client/index.js';
 export * from './acvm/acvm.js';
-export * from './abi_coder/index.js';
 export * from './public/index.js';
 export { computeSlotForMapping } from './utils.js';
