+++ yarn-project/aztec.js/src/account_impl/account_contract.ts
@@ -1,10 +1,9 @@
 import partition from 'lodash.partition';
-import { encodeArguments } from '@aztec/acir-simulator';
 import { AztecAddress, CircuitsWasm, Fr, FunctionData, TxContext } from '@aztec/circuits.js';
 import { padArrayEnd } from '@aztec/foundation/collection';
 import { sha256 } from '@aztec/foundation/crypto';
 import { ExecutionRequest, PackedArguments, PartialContractAddress, PublicKey, TxExecutionRequest } from '@aztec/types';
-import { ContractAbi, generateFunctionSelector } from '@aztec/foundation/abi';
+import { ContractAbi, encodeArguments, generateFunctionSelector } from '@aztec/foundation/abi';
 
 import { AccountImplementation } from './index.js';
 import { EcdsaAuthProvider, SchnorrAuthProvider } from '../auth/index.js';
