+++ yarn-project/end-to-end/src/utils.ts
@@ -9,7 +9,6 @@ import {
   AztecAddress,
   Contract,
   ContractDeployer,
-  Point,
   EthAddress,
   SchnorrAuthProvider,
   Wallet,
