+++ yarn-project/foundation/src/abi/abi_coder.ts
@@ -1,5 +1,5 @@
 import { keccak } from '@aztec/foundation/crypto';
-import { ABIParameter } from '@aztec/foundation/abi';
+import { ABIParameter, ABIType } from '@aztec/foundation/abi';
 
 /**
  * Generate a function signature string for a given function name and parameters.
@@ -25,4 +25,28 @@ export function generateFunctionSignature(name: string, parameters: ABIParameter
 export function generateFunctionSelector(name: string, parameters: ABIParameter[]) {
   const signature = generateFunctionSignature(name, parameters);
   return keccak(Buffer.from(signature)).slice(0, 4);
+}
+
+/**
+ * Get the size of an ABI type in field elements.
+ * @param type - The ABI type.
+ * @returns The size of the type in field elements.
+ */
+export function sizeOfType(type: ABIType): number {
+  switch (type.kind) {
+    case 'field':
+    case 'boolean':
+    case 'integer':
+      return 1;
+    case 'string':
+      return type.length;
+    case 'array':
+      return type.length * sizeOfType(type.type);
+    case 'struct':
+      return type.fields.reduce((sum, field) => sum + sizeOfType(field.type), 0);
+    default: {
+      const exhaustiveCheck: never = type;
+      throw new Error(`Unhandled abi type: ${exhaustiveCheck}`);
+    }
+  }
 }
