+++ yarn-project/foundation/src/abi/index.ts
@@ -1,4 +1,6 @@
 export * from './abi_coder.js';
+export * from './encoder.js';
+export * from './decoder.js';
 
 /**
  * A named type.
